Split SubjectRepository into model
